RavenDB-18261 - refactoring test
